regen: unblock unified backend stmt processing
This unblocks unification of backend module-level statement processing,
see pull-request: https://github.com/nim-works/nimskull/pull/714.

The outgoing csources compiler is unable to do the first bootstrap
iteration with `--gc:orc` and requires `--gc:refc`, in turn the latest
code is triggering a `refc` bug requiring a newer compiler.

On the upside, and as a requisite, `--gc:refc` will no longer be a part
of the bootstrap process.

Used c_sources from: 0.1.0-dev.20827